Fix v8-devel release field
The current approach would result in reusing the same ENVR for
v8-devel any time the nodejs release field was reused and the
v8 version provided did not change. This is the same approach
we use for NPM to avoid this problem.

We need to bump the v8-devel epoch because
v8-devel-7.5.288.22-2.fc31 > v8-devel-7.5.288.22-1.12.7.0.1.fc31

Signed-off-by: Stephen Gallagher <sgallagh@redhat.com>